+++ handlers/dashboard.go
@@ -130,25 +130,14 @@ func apiThemeCreateHandler(w http.ResponseWriter, r *http.Request) {
 	}
 	utils.Log.Infof("creating assets in folder: %s/%s", dir, "assets")
 	if err := source.CreateAllAssets(dir); err != nil {
-		log.Errorln(err)
-		sendErrorJson(err, w, r)
-		return
-	}
-	if err := source.CompileSASS(); err != nil {
 		if err := source.CopyToPublic(source.TmplBox, "css", "style.css"); err != nil {
 			log.Errorln(err)
 			sendErrorJson(err, w, r)
 			return
+		} else {
+			log.Errorln(err)
+			sendErrorJson(err, w, r)
 		}
-		jsFiles := []string{"bundle.js", "main.chunk.js", "polyfill.chunk.js", "style.chunk.js"}
-		for _, f := range jsFiles {
-			if err := source.CopyToPublic(source.TmplBox, "js", f); err != nil {
-				log.Errorln(err)
-				sendErrorJson(err, w, r)
-				return
-			}
-		}
-		log.Errorln("Default 'base.css' was inserted because SASS did not work.")
 	}
 	resetRouter()
 	sendJsonAction(dir+"/assets", "created", w, r)