+++ source/source.go
@@ -13,7 +13,7 @@ import (
 
 var (
 	log           = utils.Log.WithField("type", "source")
-	TmplBox       *rice.Box // HTML and other small files from the 'source/tmpl' directory, this will be loaded into '/assets'
+	TmplBox       *rice.Box // HTML and other small files from the 'source/dist' directory, this will be loaded into '/assets'
 	RequiredFiles = []string{
 		"css/style.css",
 		"css/style.css.gz",
@@ -25,17 +25,13 @@ var (
 		"scss/mixin.scss",
 		"scss/mobile.scss",
 		"scss/variables.scss",
-		"js/bundle.js",
-		"js/main.chunk.js",
-		"js/polyfill.chunk.js",
-		"js/style.chunk.js",
 		"banner.png",
 		"favicon.ico",
 		"robots.txt",
 	}
 )
 
-// Assets will load the Rice boxes containing the CSS, SCSS, JS, and HTML files.
+// Assets will load the Rice boxes containing the CSS, SCSS, favicon, and HTML files.
 func Assets() error {
 	if utils.Params.GetBool("DISABLE_HTTP") {
 		return nil
@@ -130,15 +126,12 @@ func OpenAsset(path string) string {
 	return data
 }
 
-// CreateAllAssets will dump HTML, CSS, SCSS, and JS assets into the '/assets' directory
+// CreateAllAssets will dump HTML, CSS, SCSS, and favicon assets into the '/assets' directory
 func CreateAllAssets(folder string) error {
 	log.Infoln(fmt.Sprintf("Dump Statping assets into %s/assets", folder))
 	fp := filepath.Join
 
 	if err := MakePublicFolder(fp(folder, "/assets")); err != nil {
-		return err
-	}
-	if err := MakePublicFolder(fp(folder, "assets", "js")); err != nil {
 		return err
 	}
 	if err := MakePublicFolder(fp(folder, "assets", "css")); err != nil {
@@ -150,7 +143,7 @@ func CreateAllAssets(folder string) error {
 	if err := MakePublicFolder(fp(folder, "assets", "scss")); err != nil {
 		return err
 	}
-	log.Infoln("Inserting scss, css, and javascript files into assets folder")
+	log.Infoln("Inserting scss, css, and favicon files into assets folder")
 
 	if err := CopyAllToPublic(TmplBox); err != nil {
 		log.Errorln(err)
@@ -199,6 +192,9 @@ func CopyAllToPublic(box *rice.Box) error {
 		if exclude[info.Name()] {
 			return nil
 		}
+		if strings.Contains(path, "js") {
+			return nil
+		}
 		if info.IsDir() {
 			return nil
 		}
@@ -232,11 +228,11 @@ func CopyToPublic(box *rice.Box, path, file string) error {
 
 // MakePublicFolder will create a new folder
 func MakePublicFolder(folder string) error {
-	log.Infoln(fmt.Sprintf("Creating folder '%v'", folder))
+	log.Infoln(fmt.Sprintf("Creating folder '%s'", folder))
 	if !utils.FolderExists(folder) {
 		err := utils.CreateDirectory(folder)
 		if err != nil {
-			log.Errorln(fmt.Sprintf("Failed to created %v directory, %v", folder, err))
+			log.Errorln(fmt.Sprintf("Failed to created %s directory, %v", folder, err))
 			return err
 		}
 	}